+++ app/Http/Controllers/DevoirController.php
@@ -16,22 +16,24 @@ class DevoirController extends Controller
 {
     /**
      * Display a listing of Devoirs
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
+        $this->authorize('view', Devoir::class);
         $devoirs = Devoir::all();
         return view('Devoirs.index',['devoirs'=>$devoirs]);
     }
 
     /**
      * Show the form for creating a new resource.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
+        $this->authorize('create', Devoir::class);
         $niveaux = Niveau::all();
         $annee = DB::table('annees')
             ->select('annees.*')
@@ -43,24 +45,26 @@ class DevoirController extends Controller
 
     /**
      * Store a newly created resource in storage.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @param  DevoirRequest  $request
      * @return \Illuminate\Http\Response
      */
     public function store(DevoirRequest $request)
     {
+        $this->authorize('create', Devoir::class);
         Devoir::create($request->all());
         return redirect()->route('devoir.index')->with('success','Devoir Ajouté');
     }
 
     /**
      * Show the form for editing the specified resource.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
+        $this->authorize('update', Devoir::class);
         $niveaux = Niveau::all();
         $annee = DB::table('annees')
             ->select('annees.*')
@@ -73,13 +77,14 @@ class DevoirController extends Controller
 
     /**
      * Update the specified resource in storage.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @param  DevoirRequest $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function update(DevoirRequest $request, $id)
     {
+        $this->authorize('update', Devoir::class);
         $devoir = Devoir::findorFail($id);
 
         $devoir->update($request->all());
@@ -89,12 +94,13 @@ class DevoirController extends Controller
 
     /**
      * Remove the specified resource from storage.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Devoir::class);
         $devoir = Devoir::findorFail($id);
         $devoir->delete();
         return redirect()->route('devoir.index')->with('success','Devoir Supprimé');
