+++ app/Http/Controllers/NoteController.php
@@ -16,9 +16,11 @@ class NoteController extends Controller
      * Display a listing of Notes
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize('view', Note::class);
         $annees = Annee::all();
         $notes = false;
         if ($user_id = $request->query('user_id')) {
@@ -29,23 +31,25 @@ class NoteController extends Controller
 
     /**
      * Show the form for creating a new resource.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
+        $this->authorize('create', Note::class);
         $annees = Annee::all();
         return view('Notes.ajout',compact('annees'));
     }
 
     /**
      * Store a newly created resource in storage.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @param  NoteRequest  $request
      * @return \Illuminate\Http\Response
      */
     public function store(NoteRequest $request)
     {
+        $this->authorize('create', Note::class);
         foreach ($request->get('notes') as $etudiantid => $note){
             if ($note) {
                 Note::create([
@@ -65,12 +69,13 @@ class NoteController extends Controller
 
     /**
      * Remove the specified resource from storage.
-     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
+        $this->authorize('create', Note::class);
         $note = Note::findorFail($id);
         $note->delete();
         return redirect()->route('note.index')->with('success','Note Supprimé');
